Integer -> Long (Int overflow is possible w/ "System.currentTimeMillis() - (days * 24 * 60 * 60 * 1000)" calculation of the searchPoint

+++ src/test/java/org/mskcc/limsrest/service/GetSequencingRequestsTaskTest.java
@@ -16,7 +16,7 @@ import static org.mockito.Mockito.when;
 public class GetSequencingRequestsTaskTest {
     @Test
     public void queriesRequestsWithFlowCellSamples() {
-        Integer numDays = 7;
+        Long numDays = 7L;
         Boolean delivered = Boolean.TRUE;
 
         GetSequencingRequestsTask task = new GetSequencingRequestsTask(numDays, delivered);
